skills/starter-django/assets/templates/requirements.txt
Django>=5.2,<5.3
django-allauth>=65.14,<66
dj-database-url>=2.3,<3
gunicorn>=23,<24
whitenoise>=6.8,<7
psycopg[binary]>=3.2,<4
python-decouple>=3.8,<4
